requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
numpy==1.26.4
pandas==2.2.3
torch==2.0.1+cu117
torchvision==0.15.2+cu117
opencv-python==4.11.0.86
matplotlib==3.10.1
seaborn==0.13.2
tqdm==4.67.1
pillow==11.2.1
scikit-learn==1.6.1
ipykernel
ttkthemes
efficientnet-pytorch
notebook